refactor(backend): Centralize Block Cost into a Single File
